# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Employee App",
        "title": "The Employee App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/employee/createEmployee": {
            "post": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Создание нового сотрудника.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Создание нового сотрудника.",
                "operationId": "createEmployee",
                "parameters": [
                    {
                        "description": "Данные нового сотрудника.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.EntityCreatingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/employee/deleteEmployee/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Удаление сотрудника.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Удаление сотрудника.",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор сотрудника.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении.",
                        "schema": {
                            "$ref": "#/definitions/transport.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/employee/getAllEmployees": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Получение списка всех сотрудников.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Получение списка всех сотрудников.",
                "operationId": "getAllEmployees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Employee"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/employee/getEmployee/{id}": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Получение данных сотрудника.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Получение данных сотрудника.",
                "operationId": "getEmployee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор сотрудника.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/employee/setDepartment": {
            "put": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Установка отдела для сотрудника.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Установка отдела для сотрудника.",
                "operationId": "setDepartment",
                "parameters": [
                    {
                        "description": "Данные установки отдела пользователю.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.SetDepartmentQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешной установке.",
                        "schema": {
                            "$ref": "#/definitions/transport.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/employee/updateEmployee/{id}": {
            "put": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Обновление данных сотрудника.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employee"
                ],
                "summary": "Обновление данных сотрудника.",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор сотрудника.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные сотрудника.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном обновлении.",
                        "schema": {
                            "$ref": "#/definitions/transport.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор отдела.",
                    "type": "integer"
                },
                "name": {
                    "description": "Наименование отдела.",
                    "type": "string"
                }
            }
        },
        "models.Employee": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Возраст сотрудника.",
                    "type": "integer"
                },
                "department": {
                    "description": "Отдел, в котором работает сотрудник.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Department"
                        }
                    ]
                },
                "id": {
                    "description": "Идентификатор сотрудника.",
                    "type": "integer"
                },
                "name": {
                    "description": "Имя сотрудника.",
                    "type": "string"
                },
                "salary": {
                    "description": "Зарплата сотрудника.",
                    "type": "integer"
                },
                "sex": {
                    "description": "Пол сотрудника.",
                    "type": "string"
                }
            }
        },
...
        "transport.EntityCreatingResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор сущности.",
                    "type": "integer"
                }
            }
        },
        "transport.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение с информацией об ошибке.",
                    "type": "string"
                }
            }
        },
        "transport.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение.",
                    "type": "string"
                }
            }
        },
        "transport.SetDepartmentQuery": {
            "type": "object",
            "properties": {
                "departmentId": {
                    "description": "Идентификатор отдела.",
                    "type": "integer"
                },
                "employeeId": {
                    "description": "Идентификатор пользователя.",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthorizationKey": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}